+++ app/src/main/java/ru/ok/technopolis/training/personal/items/ProgressItem.kt
@@ -1,7 +1,7 @@
 package ru.ok.technopolis.training.personal.items
 
 class ProgressItem(
-    val value: Float,
+    val value: Int,
     val date: String,
     val time: String? = null
 )
